drivers: uc81xx: Specify overrides in a DT child node
Introduce the concept of refresh profiles which are specified as a
child node in the device tree. This makes it possible to use different
overrides for different types of refreshes (full/partial).

The only profile that is currently supported is the "full" profile.

(cherry picked from commit 957e22976d24e8cffbb846789ebd8a126469aa92)

Original-Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>
GitOrigin-RevId: 957e22976d24e8cffbb846789ebd8a126469aa92
Change-Id: Id24493458e1027139436d5d89369aed3f5efa585
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3870786
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>