FROMPULL: cmake: remove device_extern.h logic
This file is no longer used by device.h, so let's avoid spending time
generating the content formerly in it.

In order to leave a pointer in place for users who are expecting to
see it or are pulling it into their own builds, however, replace its
contents with an #error directive that tells them what happened. This
can be removed later on when we expect people are used to the
transition.

From pull request at
https://github.com/zephyrproject-rtos/zephyr/pull/44883.

BUG=b:241260292
TEST=zmake testall

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Change-Id: I718d35cb92ecdad10192f51c6b1f8257305d969b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3805858
Auto-Submit: Abe Levkoy <alevkoy@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Abe Levkoy <alevkoy@chromium.org>
Commit-Queue: Abe Levkoy <alevkoy@chromium.org>