drivers: ssd16xx: Remove SCREEN_INFO_DOUBLE_BUFFER cap
The SSD16xx driver used to use the SCREEN_INFO_DOUBLE_BUFFER flag to
indicate to the LVGL integration that it needs writes to be performed
twice. This was required because partial writes require both the old
and new buffer to be written.

This behavior is really an implementation detail and only applies to
partial refresh. Do this buffer maintenance in the driver instead.

(cherry picked from commit 6535d6a0e66df21cbfa7d2b949763d5a74930583)

Original-Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>
GitOrigin-RevId: 6535d6a0e66df21cbfa7d2b949763d5a74930583
Change-Id: If6c399a9ed3d941bf3b4473d95bf2fa83a118feb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4526238
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>