drivers: ieee802154: rf2xx: Add support to Sub-Giga
Add support to at86rf212[b] sub-giga devices. This work enables use of
pages 0, 2 and 5 in accordance with IEEE-802.15.4/2003/2006/2011. The
proprietary speeds can be object of future work.

performance, low emissions and to save power. A reference power table
can be found in the datasheet and should be used for tests only and
not on a final product.

(cherry picked from commit 6d5cdc17b4cd3f41afdf61eceb839be88780c09e)

Original-Note: It is recommended that user define a power table for better
Original-Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>
GitOrigin-RevId: 6d5cdc17b4cd3f41afdf61eceb839be88780c09e
Change-Id: Ida6015a83f50dad9cb2420a2d3f2e52b091e20b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3889538
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>