Bluetooth: Controller: Fix directed adv connection setup stall
When directed advertising and continuous scanning are
simultaneously active, and a connection request is received
by the directed advertising state then the connection was
not established due to the done event not being generated.

Done event from the directed advertising state was not
generated because the prepare pipeline only had resume
event which does not get scheduled when there is prepare
event of the continuous scan state.

The fix is to explicitly abort all pending prepare and
resume event in the pipeline for the directed advertising
state when connection request is received.

(cherry picked from commit 3eb0c448be3abeebf4e9bdb281e776a019268051)

Original-Signed-off-by: Vinayak Kariappa Chettimada <vich@nordicsemi.no>
GitOrigin-RevId: 3eb0c448be3abeebf4e9bdb281e776a019268051
Change-Id: I7e73880bdb8ed1fe6a16dafda4204b921c3f5d49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3807706
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>