Bluetooth: host: add missing meta data destroy callback
Symptoms were that a disconnect happening when a buffer was enqueued but
not sent resulted in a meta-data memory leak.

This is problematic because it seemed (as per the OP) that it resulted in
the whole host getting locked up when a device sending long L2CAP packets
was prone to disconnect-reconnect cycles (e.g. in a bad RF environment).

Fixes #47649

(cherry picked from commit 00709e5fbbe0b91c7aa4c0c9effd8b3abb68210c)

Original-Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>
GitOrigin-RevId: 00709e5fbbe0b91c7aa4c0c9effd8b3abb68210c
Change-Id: I038aec30c304f7d86dfff0903915e661fe816940
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3804475
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>