tests: lib: notify: Fix uninitialised variable warning
This commit sets an initial value of 0 for the `res` variable, whose
pointer is passed to other functions and is not directly assigned
within the calling function.

Note that, when the test completes successfully, the value of the `res`
variable should be set to `423` (the value of `set_res`).

This fixes the "‘res’ may be used uninitialized" warning generated by
the GCC 12.

(cherry picked from commit bde270b1d5d70b3d7194cd8630188f5ea2c48665)

Original-Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: bde270b1d5d70b3d7194cd8630188f5ea2c48665
Change-Id: Ic56c714e35bd129e37032f0aa635a61912f7ca41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3795403
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>