samples: zbus: Add work queue sample
The work_queue sample illustrates three reaction styles available for
using zbus. The first is a listener that reacts by callback; use it should
for urgent reaction. The second is a listener that responds by callback;
instead of executing the code, it pushes a job to a work queue that will
be executed shortly but not immediately. The last one is the subscriber
that reacts using a queue; use it for an asynchronous reaction where the
developer would like to control the flow.

(cherry picked from commit 5ed4781c54c67f62eae3e1d8514aeddfa70b72cb)

Original-Signed-off-by: Rodrigo Peixoto <rodrigopex@gmail.com>
GitOrigin-RevId: 5ed4781c54c67f62eae3e1d8514aeddfa70b72cb
Change-Id: I9f583be083f591f026bc9aff0c918e65aed4416f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4028202
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>