i2c: shell: Remove prefix filtering of device list
The I2C shell would filter the list of devices it knows about to
ones that have a device name that starts with "I2C_".  It was the
case that the majority of I2C bus controller devices happened
to be named with the "I2C_" prefix, however there is no guarantee
that is the case.  With the remove of label properties from the
devicetree this is even more true.

For now remove the prefix filter and just return the full list
of devices.

(cherry picked from commit 5a4d1f7ac3acd83774c4761ace6a718080fc1846)

Original-Signed-off-by: Kumar Gala <galak@kernel.org>
GitOrigin-RevId: 5a4d1f7ac3acd83774c4761ace6a718080fc1846
Change-Id: I056a5a281dfdde040aa46f645de89247d303dca4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3816323
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>