boards: nrf9160dk_nrf9160: Correct arduino_spi definition
This is a follow-up to commit 854129902057ad236acc8ff3b091ededc86f922c.

The arduino_spi definition introduced in the above commit is incorrect
because the spi1 MOSI and MISO pins are swapped and spi1 uses the same
pins (P0.13, P0.12, and P0.11) that are assigned to spi3 when building
for board revision 0.14.0 and above.
This commit actually reverts the changes made to dts files in the
commit mentioned above and instead it adds the arduino_spi label
to spi3 and assigns the proper pins to this instance in the default
dts file. And in the overlay for revision 0.14.0 it just adds another
CS line to spi3 (the SPIM pins are by default configured correctly)
so that this instance can be used for communication with the exernal
flash.

(cherry picked from commit af75354d8206f40ead464fff4f0a24552e8aed23)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: af75354d8206f40ead464fff4f0a24552e8aed23
Change-Id: If3640545653dd4601351448c366d5c39f51a2098
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3819564
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>