lib: os: assert: Add unreachable path hint for assertion failure
This commit adds the `CODE_UNREACHABLE` hint at the end of the
assertion failure branch so that the compiler takes note of the assert
function not returning when an assertion fails.

This prevents the compiler from generating misguided warnings assuming
the asserted execution paths.

It also introduces the `ASSERT_TEST` Kconfig symbol, which indicates
that the "assert test mode" is enabled. This symbol may be selected by
the tests that require the assert post action function to return
without aborting so that the test can proceed.

Note that the `CODE_UNREACHABLE` hint is specified only when the assert
test mode is disabled in order to prevent the tests from crashing when
the assert post action function returns.

(cherry picked from commit 230ddd9a7f91194be762d24a3ee5d4c011ee2467)

Original-Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: 230ddd9a7f91194be762d24a3ee5d4c011ee2467
Change-Id: Iffee3c2f65a63b99f0050e06f7725e340e18579f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3826646
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>