drivers: timer: nrf_rtc_timer: Fix assert conditions
This commit fixes the incomplete assert conditions for the `chan`
argument passed to the nRF RTC timer functions.

Note that the `chan` argument for this driver is of a **signed**
integer type, so it is necessary to check that its value is
non-negative.

This fixes the warnings generated by the GCC 12 such as:

  error: array subscript -1 is below array bounds of '...'

(cherry picked from commit 4004475e7fbc93f9e845eca7d4bdf36599260d6c)

Original-Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: 4004475e7fbc93f9e845eca7d4bdf36599260d6c
Change-Id: I71d8f82115467822481a1ec12fe494765bcc5d34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3814834
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>