soc: arm: stm32g0: fix USB-C PD dead battery support
Currently the USB-C PD dead battery support is disabled very early in
the boot process, in order to give access to CC1 and CC2 pins to the
application. However this breaks dead battery support as this instructs
the PD source to stop delivering power.

This commit changes the behavour to keep the internal pull-up in dead
battery pins enable if the UCPD driver is enabled, and the corresponding
UCPD peripheral is also enabled. The code is a bit complex as it aims to
support possible G0 devices with a single UCPD controller (as opposed to
the current one), and also because I haven't found any simple way to
match the UCPD peripheral with the bits in the SYSCFG_CFGR1 register (I
used the same trick with the base address as in the driver).

Given I have not been able to get that working with macros, the check is
done at "runtime", however GCC is smart enough to optimize the code. The
generated code is identical if none of the UCPD peripheral is enabled.

(cherry picked from commit 72dab5bda9f3396e62f213b513dfa75f6e7620d5)

Original-Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>
GitOrigin-RevId: 72dab5bda9f3396e62f213b513dfa75f6e7620d5
Change-Id: I23e6928eb410133908e9a04478083c1d0da6eb93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3831895
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>