wifi_mgmt: Add support for frequency bands
Channels are not unique across bands, they need to be interpreted along
with frequency band. Add a separate band parameter that supports 2.4, 5
and 6 GHz bands.

Also, 6GHz has maximum channel of 233 (as of now), so, change the
maximum channel to that, this is a global maximum, per-band channel
validation is left to the offload driver as that depends on regulatory
rules (not a simple maximum check).

(cherry picked from commit dace99ebdd0c07b2089d4da0dd290f8dd6dfa188)

Original-Signed-off-by: Krishna T <krishna.t@nordicsemi.no>
GitOrigin-RevId: dace99ebdd0c07b2089d4da0dd290f8dd6dfa188
Change-Id: Ic0ea29348c684caaba0e817397ac7e97c0b1b4e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3826653
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>