libc/picolibc: Place malloc heap in noinit section if possible
When the heap is of a fixed size and there isn't a special malloc partition
in use, place the heap in uninitialized memory so that the application
doesn't spend time at startup erasing it. Picolibc malloc always clears
memory before returning it to applications, so this change will not be
visible to applications.

(cherry picked from commit f0216a2619da7f5078976918a3566b9dab204589)

Original-Signed-off-by: Keith Packard <keithp@keithp.com>
GitOrigin-RevId: f0216a2619da7f5078976918a3566b9dab204589
Change-Id: I660274677256f68db6b331c9d1d7b73debe9e9e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3826696
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>