drivers: sensor: adxl372: Fix default bus issue
The adxl372 device does not select the correct bus that it is on, which
causes undefined operation (usually a fault). This driver should not be
using Kconfig to determine which bus to use, that should be done by
dts, however this works as a quick fix in the interim until the driver
can be fixed properly.

(cherry picked from commit 9433555f0d862cfb5bd1133ff1f5a5cb19cb6f6e)

Original-Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>
GitOrigin-RevId: 9433555f0d862cfb5bd1133ff1f5a5cb19cb6f6e
Change-Id: I2049e914e04a476882803ca5cac1e335b77acc28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3830615
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>