lib: os: cbprintf: move padding for xtensa from hdr_ext to hdr
This moves padding needed for Xtensa in the cbprintf package header
from the extended header to the simple header. This is due to
the cbprintf packaging function uses the simple header to calculate
offset. This results in the padding not being there in the package,
and the printing function reading incorrect data in the package.

This was discovered while looking into #49016. Without the padding
in the simple header, turning on tagged argument on cbprintf would
crash on qemu_xtensa.

(cherry picked from commit 203c4c2e13ff0213d28c9bbf365e2dc71fb2230e)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: 203c4c2e13ff0213d28c9bbf365e2dc71fb2230e
Change-Id: I13daa65e3a248f6e9b348414515f8d577a1d2b29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3879794
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Aaron Massey <aaronmassey@google.com>