boards: arm: lpcxpresso55s69: Fix build of lpcxpresso55s69_cpu1
The samples/sensor/thermometer fails to build on lpcxpresso55s69_cpu1
due to the fact that we try and build the fxos8700 driver however the
bus that this driver is on is not enabled.  Disable the sensor in the
devicetree since the bus is not enabled.

(cherry picked from commit f2a6c738002299f97d463aa14d3406334514c31f)

Original-Signed-off-by: Kumar Gala <galak@kernel.org>
GitOrigin-RevId: f2a6c738002299f97d463aa14d3406334514c31f
Change-Id: Ie3e936ae76a6d1dc1e9d32e8d9b603292f100819
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3838035
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>