drivers: pinctrl_nrf: Fix nrf_pin_configure() implementation
This is a follow-up to commit fd0767557439d04c4a220a9857e901d8d71498db.

The above commit was supposed to introduce overriding of the S0S1
drive setting with S0D1 for TWI/TWIM peripherals, but since it did
not properly update the `nrf_pin_configure()` function (the `drive`
parameter was only added in the function signature, but then it was
not used...), the drive setting was in fact not overridden.
This commit corrects this embarrassing oversight.

(cherry picked from commit 32a87215d7aa97191097357f602237f7da4a3536)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 32a87215d7aa97191097357f602237f7da4a3536
Change-Id: Ie2b5180b2b014e2703c261f22755fe3012880b96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3849675
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>