tests: Increase stack size when nRF DK serial workaround is enabled
... and tests are built with no optimizations. The standard size may
not be sufficient in such circumstances.

Commit 4b1103547cb8c3acd622ea8c1540e592eeb49c02 addressed the same
issue in a single test and used a higher value (1024) that turns out
to be not necessary, so revert also that change to make the cleanup
easier when the workaround will no longer be needed.

(cherry picked from commit 326f7bd45056023b62bc7c382ebebb14205dbca1)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 326f7bd45056023b62bc7c382ebebb14205dbca1
Change-Id: I67bc0c65a71621dcca20bd855d96efba9589a9bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3882685
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>