twister: improve logging of cmake commands
With verbose output enabled, twister logs the cmake arguments as a
python array whenever it invokes cmake to generate a build system.

This is clunky output for manual reproduction, because it forces the
developer to manually join the arguments together for consumption by
the shell.

As an enhancement for this use case, use the standard library
shlex.join() function to produce a string that can be copy/pasted
directly into the shell on POSIX platforms.

We should use this function instead of str.join, because it correctly
handles things like shell quoting for options with spaces in them.

This function is not available on Windows, so we fall back on behavior
that is similar to the old one there.

The main difference on Windows is that the output now includes the
path to cmake as well, and not just its arguments. (This is the same
on POSIX, and is intended to help debug if multiple cmakes are
installed on the same machine.)

We are running cmake from a couple of different places, so doing this
cleanly requires adding a shared module with the platform abstraction.

(cherry picked from commit 88aa873fbc544b99c1e48e37d5c74662f90b3de0)

Original-Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: 88aa873fbc544b99c1e48e37d5c74662f90b3de0
Change-Id: Ief79fa8de512a867d97210f080d4a4617e799013
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3859472
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: Aaron Massey <aaronmassey@google.com>