dts: usb-c: Introduce usb-c-connector
A USB TypeC connector has many peripherals associated with
it and the DTS binding in this commit provides a way to group
peripherals and properties in a device tree.
This binding is used with the USB-C Subsytem.

This is based on Linux, documentation:
https://www.kernel.org/doc/Documentation/devicetree/bindings/connector/usb-connector.yaml

(cherry picked from commit a1e6e9dfbe8183626ad05156668bd35d02e9f7a6)

Original-Signed-off-by: Sam Hurst <sbh1187@gmail.com>
GitOrigin-RevId: a1e6e9dfbe8183626ad05156668bd35d02e9f7a6
Change-Id: I10e0377475e10fc6769f8a12e31a54a657b137dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3979579
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>