twister: Remove import exception handler for psutil
`handlers.py` treats the `psutil` package as optional and quietly prints
a message to stdout if it is not present, but it is actually a hard
requirement for the base `Handler` class. This try/except block is bad
as it hides the requirement until a timeout occurs and the handler
attempts to use `psutil` in the terminate method, which will crash if
the module happens to be missing.

This PR removes the try/except guard so missing `psutil` will cause
twister to immediately fail. `psutil` is already specified in the
`requirements.txt` file, so this should only affect users who
unwittingly have an outdated/incorrect environment.

Also update the Github Action for twister tests to install from
requirements files.

(cherry picked from commit 52db2d8279005515907626467ab82c3ec40f24c6)

Original-Signed-off-by: Tristan Honscheid <honscheid@google.com>
GitOrigin-RevId: 52db2d8279005515907626467ab82c3ec40f24c6
Change-Id: I307a751d28417739543793f4a23f5ff7ca2a9f42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3869143
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>