Bluetooth: Host: Add `CONFIG_BT_ID_UNPAIR_MATCHING_BONDS`
This commit prevents the assertion error added in the previous commit,
by introducing the following behavior.

When the Host detects that a pairing procedue would result in more than
one local identity having a bond with the same remote address, it will
try to abort the pairing. If the pairing procedure cannot be properly
aborted, it will remove the pairing locally and disconnect.

This commit also introduces a new kconfig `BT_ID_UNPAIR_MATCHING_BONDS`.
This config changes the above-described behavior to instead un-pair the
old bond and continue with the pairing (if the new bond has the same or
better security.)

The new kconfig is not enabled by default. See the help text of the new
kconfig option for more details.

(cherry picked from commit fd5fe6ca8e8fbc89263ad16f78403989e1308870)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: fd5fe6ca8e8fbc89263ad16f78403989e1308870
Change-Id: I369132d2fb30ba15304aefdc342bca840bba3f69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3988051
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>