mcumgr: img_mgmt: Add SHA256 image upload hash check
Will check an uploaded image's hash against the hash that was
originally supplied before the upload began and return the result to
the client to know if the upload was successful or if there was an
error during upload. Requires CONFIG_IMG_ENABLE_IMAGE_CHECK be
enabled for functionality to be available.

(cherry picked from commit 23ca3841db56fc0326ceac26a0947897beab932a)

Original-Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>
GitOrigin-RevId: 23ca3841db56fc0326ceac26a0947897beab932a
Change-Id: Idcde0c11b6b9c66c6f5726c588bb143298e15b43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4122547
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>